undoes all changes we had introduced to HiGHS and updates the ipx code to the latest from the ergo-code ipx repo

+++ src/ipm/ipx/src/diagonal_precond.cc
@@ -55,8 +55,8 @@ void DiagonalPrecond::_Apply(const Vector& rhs, Vector& lhs,
     Timer timer;
 
     assert(factorized_);
-    assert((int)lhs.size() == m);
-    assert((int)rhs.size() == m);
+    assert(lhs.size() == m);
+    assert(rhs.size() == m);
 
     for (Int i = 0; i < m; i++) {
         lhs[i] = rhs[i] / diagonal_[i];
